mv cgi func to process_client_event.cpp
