+++ srcs/HttpResponse/HttpResponse.cpp
@@ -182,164 +182,6 @@ bool HttpResponse::is_status_error() const {
 }
 
 
-bool HttpResponse::is_exec_cgi() {
-    if (this->request_.method() != kGET && this->request_.method() != kPOST) {
-        return false;
-    }
-    std::pair<ScriptPath, PathInfo> pair = get_script_path_and_path_info();
-    std::string script_path = pair.first;
-    Result<bool, StatusCode> result = FileHandler::is_file(script_path);
-    return result.is_ok();
-}
-
-
-/*
- path/to/script.cgi/path/info
-                    ^^^^^^^^^ PATH_INFO
-
-  PATH_INFO = "" | ( "/" path )
-  path      = lsegment *( "/" lsegment )
-  lsegment  = *lchar
-  lchar     = <any TEXT or CTL except "/">
-  TEXT      = <any printable character>
-  CTL       = <any control character>
-  https://tex2e.github.io/rfc-translater/html/rfc3875.html#4-1-5--PATHINFO
- */
-std::pair<ScriptPath, PathInfo> HttpResponse::get_script_path_and_path_info() {
-    std::string target = this->request_.target();
-    std::string script_path, path_info;
-
-    DEBUG_PRINT(MAGENTA, "get script_path and path_info");
-    std::size_t slash_pos = 0;
-    while (slash_pos < target.length()) {
-        slash_pos = target.find('/', slash_pos);
-        if (slash_pos == std::string::npos) {
-            break;
-        }
-        std::string tmp_script_path = target.substr(0, slash_pos);
-        if (Config::is_cgi_extension(this->server_config_, tmp_script_path)) {
-            script_path = tmp_script_path;
-            path_info = target.substr(slash_pos + 1);
-            break;
-        }
-        ++slash_pos;
-    }
-
-    if (script_path.empty() && Config::is_cgi_extension(this->server_config_, target)) {
-        script_path = target;
-    }
-
-    std::string root;
-    Result<std::string, int> root_result = Config::get_root(this->server_config_,
-                                                            script_path);
-    if (root_result.is_ok()) {
-        root = root_result.ok_value();
-        if (!root.empty() && root[root.length() - 1] == '/' && script_path[0] == '/') {
-            script_path = script_path.substr(1);
-        }
-        script_path = root + script_path;
-    }
-
-    DEBUG_PRINT(MAGENTA, " script_path and path_info");
-    DEBUG_PRINT(MAGENTA, "  script_path[%s]", script_path.c_str());
-    DEBUG_PRINT(MAGENTA, "  path_info  [%s]", path_info.c_str());
-    return std::make_pair(script_path, path_info);
-}
-
-
-CgiParams HttpResponse::get_cgi_params(const std::string &script_path,
-                                       const std::string &path_info) {
-    CgiParams params;
-
-    if (this->request_.method() == kPOST) {
-        params.content = this->request_.body();
-        params.content_length = params.content.size();
-        params.content_type = this->request_.content_type();
-    }
-    params.query_string = this->request_.query_string();
-    params.path_info = path_info;
-    params.script_path = script_path;
-
-    DEBUG_PRINT(MAGENTA, "cgi params ");
-    DEBUG_PRINT(MAGENTA, " content       : [%s]", std::string(params.content.begin(), params.content.end()).c_str());
-    DEBUG_PRINT(MAGENTA, " content_length: [%zu]", params.content_length);
-    DEBUG_PRINT(MAGENTA, " content_type  : [%s]", params.content_type.c_str());
-    DEBUG_PRINT(MAGENTA, " query_string  : [%s]", params.query_string.c_str());
-    DEBUG_PRINT(MAGENTA, " path_info     : [%s]", params.path_info.c_str());
-    DEBUG_PRINT(MAGENTA, " script_path   : [%s]", params.script_path.c_str());
-
-    return params;
-}
-
-
-ProcResult HttpResponse::exec_cgi_process() {
-    std::pair<ScriptPath, PathInfo> pair = HttpResponse::get_script_path_and_path_info();
-
-    CgiParams params = get_cgi_params(pair.first, pair.second);
-    this->cgi_handler_.set_cgi_params(params);
-
-    if (this->cgi_handler_.exec_script(params.script_path) == Failure) {
-        this->set_status_code(InternalServerError);
-        this->clear_cgi();
-        return Failure;
-    }
-    return Success;
-}
-
-
-/*
- 6.2.1. Document Response
- document-response = Content-Type [ Status ] *other-field NL response-body
-
- The script MUST return a Content-Type header field.
- https://tex2e.github.io/rfc-translater/html/rfc3875.html#6-2-1--Document-Response
-
- Content-Type = "Content-Type:" media-type NL
- https://tex2e.github.io/rfc-translater/html/rfc3875.html#6-3-1--Content-Type
-
- Status         = "Status:" status-code SP reason-phrase NL
- status-code    = "200" | "302" | "400" | "501" | extension-code
- extension-code = 3digit
- reason-phrase  = *TEXT
- https://tex2e.github.io/rfc-translater/html/rfc3875.html#6-3-3--Status
-
- other-field     = protocol-field | extension-field
- protocol-field  = generic-field
- extension-field = generic-field
- generic-field   = field-name ":" [ field-value ] NL
- field-name      = token
- field-value     = *( field-content | LWSP )
- field-content   = *( token | separator | quoted-string )
- https://tex2e.github.io/rfc-translater/html/rfc3875.html#6-3--Response-Header-Fields
-
- response-body = *OCTET
- https://tex2e.github.io/rfc-translater/html/rfc3875.html#6-4--Response-Message-Body
-
- UNIX
- The newline (NL) sequence is LF; servers should also accept CR LF as a newline.
-                                          ^^^^^^
- */
-void HttpResponse::interpret_cgi_output() {
-    if (this->status_code() != StatusOk) {  // todo: cgi_status ??
-        return;
-    }
-
-    StatusCode parse_status = this->cgi_handler_.parse_document_response();
-    this->set_status_code(parse_status);
-
-    if (is_status_error()) {
-        return;
-    }
-    if (!is_supported_by_media_type(this->cgi_handler_.content_type())) {
-        this->set_status_code(NotAcceptable);
-        return;
-    }
-
-    this->body_buf_ = this->cgi_handler_.cgi_body();
-    add_content_header_by_media_type(this->cgi_handler_.content_type());
-}
-
-
 ProcResult HttpResponse::send_http_response(int client_fd) {
     return Socket::send_buf(client_fd, &this->response_msg_);
 }
